fix-ios-build

+++ mobile/modules/core/ios/Packages/CoreObjC/PcmConverter.h
@@ -11,7 +11,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface PcmConverter : NSObject
 + (void)setupStaticEncoderAndDecoder;
--(NSMutableData *)decode: (NSData *)lc3data frameSize:(uint16_t)frameSize;
+-(NSMutableData *)decode: (NSData *)lc3data frameSize:(NSInteger)frameSize;
 -(NSMutableData *)encode: (NSData *)pcmdata;
 @end
